Modified the database fetch operation to no longer use react hooks, now it should run on the server side, so the fetch operation does no occur everytime the website refreshes.
